Add dto

+++ src/main/java/com/example/vuelos/controllers/ReservaController.java
@@ -1,6 +1,8 @@
 package com.example.vuelos.controllers;
 
-import com.example.vuelos.dtos.ReservaDTO;
+import com.example.vuelos.controllers.dtos.ReservaDTO;
+import com.example.vuelos.controllers.dtos.ReservaRequestDTO;
+import com.example.vuelos.exceptions.ResourceNotFound;
 import com.example.vuelos.services.ReservaService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,25 +25,26 @@ public class ReservaController {
 
     @GetMapping("/{id}")
     public ResponseEntity<ReservaDTO> getReservaById(@PathVariable Long id) {
-        return reservaService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.ok(reservaService.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Reserva no encontrada con id: " + id)));
     }
 
     @PostMapping
-    public ResponseEntity<ReservaDTO> createReserva(@RequestBody ReservaDTO reservaDTO) {
-        return ResponseEntity.ok(reservaService.create(reservaDTO));
+    public ResponseEntity<ReservaDTO> createReserva(@RequestBody ReservaRequestDTO reservaRequestDTO) {
+        return ResponseEntity.ok(reservaService.create(reservaRequestDTO));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ReservaDTO> updateReserva(@RequestBody ReservaDTO reservaToUpdateDTO, @PathVariable Long id) {
-        return reservaService.update(id, reservaToUpdateDTO)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<ReservaDTO> updateReserva(@RequestBody ReservaRequestDTO reservaRequestDTO, @PathVariable Long id) {
+        return ResponseEntity.ok(reservaService.update(id, reservaRequestDTO)
+                .orElseThrow(() -> new ResourceNotFound("Reserva no encontrada con id: " + id)));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReserva(@PathVariable Long id) {
+        if (!reservaService.findById(id).isPresent()) {
+            throw new ResourceNotFound("Reserva no encontrada con id: " + id);
+        }
         reservaService.delete(id);
         return ResponseEntity.noContent().build();
     }
